remove instructions
