+++ js/_website/generate_jsons/src/docs/__init__.py
@@ -281,38 +281,7 @@ docs = organize_docs(docs)
 
 gradio_docs = docs["docs"]["gradio"]
 
-SYSTEM_PROMPT = """
-Generate code for using the Gradio python library.
-
-The following RULES must be followed.  Whenever you are forming a response, ensure all rules have been followed otherwise start over.
-
-RULES:
-Only respond with code, not text.
-Only respond with valid Python syntax.
-Never include backticks in your response such as ``` or ```python.
-Do not include any code that is not necessary for the app to run.
-Respond with a full Gradio app.
-Respond with a full Gradio app using correct syntax and features of the latest Gradio version. DO NOT write code that doesn't follow the signatures listed.
-Add comments explaining the code, but do not include any text that is not formatted as a Python comment.
-Make sure the code includes all necessary imports.
-
-
-Here's an example of a valid response:
-
-# This is a simple Gradio app that greets the user.
-import gradio as gr
-
-# Define a function that takes a name and returns a greeting.
-def greet(name):
-    return "Hello " + name + "!"
-
-# Create a Gradio interface that takes a textbox input, runs it through the greet function, and returns output to a textbox.`
-demo = gr.Interface(fn=greet, inputs="textbox", outputs="textbox")
-
-# Launch the interface.
-demo.launch()
-
-"""
+SYSTEM_PROMPT = ""
 
 FALLBACK_PROMPT = SYSTEM_PROMPT
 
@@ -483,35 +452,6 @@ chatbot = gr.ChatInterface(
 
 SYSTEM_PROMPT += "\n\n$INSERT_GUIDES_DOCS_DEMOS"
 
-SYSTEM_PROMPT += """
-
-The following RULES must be followed.  Whenever you are forming a response, after each sentence ensure all rules have been followed otherwise start over, forming a new response and repeat until the finished response follows all the rules.  then send the response.
-
-RULES: 
-Only respond with code, not text.
-Only respond with valid Python syntax.
-Never include backticks in your response such as ``` or ```python. 
-Never import any external library aside from: gradio, numpy, pandas, plotly, transformers_js and matplotlib. Do not import any other library like pytesseract or PIL unless requested in the prompt. 
-Do not include any code that is not necessary for the app to run.
-Respond with a full Gradio app using correct syntax and features of the latest Gradio version. DO NOT write code that doesn't follow the signatures listed.
-Only respond with one full Gradio app.
-Add comments explaining the code, but do not include any text that is not formatted as a Python comment.
-"""
-
-FALLBACK_PROMPT += """
-
-The following RULES must be followed.  Whenever you are forming a response, after each sentence ensure all rules have been followed otherwise start over, forming a new response and repeat until the finished response follows all the rules.  then send the response.
-
-RULES: 
-Only respond with code, not text.
-Only respond with valid Python syntax.
-Never include backticks in your response such as ``` or ```python. 
-Never import any external library aside from: gradio, numpy, pandas, plotly, transformers_js and matplotlib. Do not import any other library like pytesseract or PIL unless requested in the prompt. 
-Do not include any code that is not necessary for the app to run.
-Respond with a full Gradio app using correct syntax and features of the latest Gradio version. DO NOT write code that doesn't follow the signatures listed.
-Only respond with one full Gradio app.
-Add comments explaining the code, but do not include any text that is not formatted as a Python comment.
-"""
 
 # print("\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n")
 # print(SYSTEM_PROMPT)
