+++ js/_website/src/routes/llms.txt/+server.ts
@@ -4,8 +4,7 @@ import SYSTEM_PROMPT from "$lib/json/system_prompt.json";
 export const prerender = true;
 
 export async function GET({ url }) {
-	const worker_url = "https://llms-txt.playground-worker.pages.dev/api/prompt";
-	// const worker_url = "https://playground-worker.pages.dev/api/prompt";
+	const worker_url = "https://playground-worker.pages.dev/api/prompt";
 	// const worker_url = "http://localhost:5173/api/prompt";
 
 	// const query = url.searchParams.get("q")?.toLowerCase() || "";
